Version 1 of database save & migration tool
Migration tool :
The migration tool is based on the project the appwrite.json is set. It can :
- get all collection from a given database
- get all buckets from the project
- create a database in the project with all collection and buckets needed.
- generate a environments.ts file for the frontend app

Database save :
the collections.json & storage.json contains all the informations. the commands.bat contains all the commands generated by the migration tool to upload the data.
You can upload the database save with the migration tool

Change the project ID, so this repo is ready for public